playing around with prefix sums
The problem right now is that the prefix sum of the light partition is
not monotone.

My assumption right now is that because i was trying to be smart and
precompute the prefix sum over all weights before hand because it
doesn't really come with a significant speed hit if we have to compute a
reduction anyway. The problem i believe comes by computing the weight
prefix sum seperately we get numerical instabilities when comparing the
two.