Finally the prefix sums over the partitions are monotone.
I decided to make this optional, because it comes with a slight
performance hit.
I tested with 1024 * 2048 weights and the latency increased by around
0.2 - 0.3ms, which may be acceptable time will tell.
A alternative would be to adjust the next step accordingly,
which may be more optimal if the binary search is bound by memory
such that these checks don't effect performance there.

The problem was due to the non associative floating point additions.
During the look back for partitioning and prefix sum calculation the
amount of lookback performed is non determinisitc, which leads
to cases where batches of weights (called partitions in the original
paper), where added inconsistatly (e.g (a+b)+c != a + (b +c)),
which lead to the monotonic invariant beeing broken. As floating point
operations do not provide such guarantees.
Further in the following computation of the prefix sums of
a single batch similar problems occured when adding the large aggregates
of previous batches to the local prefix sums, which meant a large number
was added to a really small one. This is especially notable for the
light partition prefix sum.

Numeric stability however still remains uneffected by the changes,
the worst case error from our reference calculations on the cpu with
doubles are around 0.5 - 1.5, with around 2 million weights, however we
have not tested if the error is distributed evenly acros the prefix
sums. If a single outlier is responsible for the numeric error this may
show itself to be problematic as this item will be drastically
oversampled.

At this stage we can most likely move on to the next step of alias table
constructions, even thou further optimizations are definetly possible.
- We currently don't use any subgroup instructions, we believe that
  shared memory footprints can be reduced by a factor of
  gl_SubgroupSize.
- We did not parallize the decoupled lookback, the original paper
  of the decoupled lookback prefix scan proposes this as a optimization technique.
- It might be possible to not write the light and heavy partitions back
  to global memory, as all following steps of alias table construction does not
  require the partitions directly it only requires the prefix sums over
  the partitions, this reduces the memory bound significatly and
  preliminary tests show a significant speedup.