add is_alias_table assertion, which checks invariants.
The current implementation expects a normalized alias table, meaning
that the probabilty or redirection is stored as a actual probabilty from
this we can reconstruct the relative probabilty of all elements,
which we compare against the weight set.

Further this commit implements a alternative reduction method that we
found in the https://github.com/lorenzhs/wrs repo of the "Parallel
Weighted Random Sampling" paper my Lorenz Hübschle-Scheider et.al.
It's a block wise reduction, which accumulates blockSize elements and
then recursivly combines blocks. However it does not improve the
numerical stability compared to our tree-reduction implementation,
therefor, we will not utilize it during testing.

Further we implemented the synchronous sweeping alias table construction
sourced from the same paper, (original paper by Vose). To test our
alias table assertions, however we encountered numerical instabilities,
because the sweeping construction accumulates a error, which is then
dumped on the last element. We see a significant error at around 1e4
elements, we want to work further on testing by implementing a reference
algorithm of the PSA algorithm to check if it will improve stability.
Our hypothesis here is that PSA might increase the stability because
errors can't accumulate over splits, but we will see if that is actually
true.