Implements the pack method of the PSA algorithm on the CPU
This was a long road to get this to work propertly.
We started by simply implementing the algorithm with the refernece to
the original paper by Sanders (https://arxiv.org/pdf/1903.00227),
which was suprisingly difficult as the paper uses a one based indexing
which was a bit confusing to convert as ranges defined by the split step
are not simply inclusive or exclusive. Further the paper does not
mention certain edge cases that have to be made due to numerical
instabilities, which often lead to entries in the alias table not beeing
packed properly or not packed at all. Further we found it quite
difficult to argue about the numerical instabiltiies because some of the
steps are we made the in current implementation do not seem completely
correct, but after looking at the implementaiton of the
sweeping construciton by Sanders (https://github.com/lorenzhs/wrs) we
saw that they made similar steps. Another paper by Lehman (https://arxiv.org/pdf/2106.12270)
actually uses similar approaches (source code at https://github.com/ByteHamster/alias-table-gpu.git)

Now verifying this was pretty hard initally we simply tested the
invariants of the constructed alias table, where we found that the
sampled weights after construction where sometimes off by a lot. However a
absolute difference in the weights does not suggest that our sampling is
biased as further analysis showed. By computing the loglog RMSE curve
for our PSA implementation we found that it converges correctly. We did
further checks to ensure that our calculations of the RMSE curve is
correct by also computing the loglog RMSE curves for the
std::discrete_distribution, which we assume to be correct. Here we had
some problems getting this to work and some pretty obvious bugs that
took ages to find. But in the end we saw 3 similar graphs for our PSA
implementation, the sweeping alias table construction by Vose and the
std::discrete_distribution, which makes us believe that we implemented a
unbiased sampler.

During testing and evaluation a lot of considerations where made to
numerical stability therefor we added numerically more stable prefix sum
and reduction functions that use the kahan summation algorithm to
improve floating point instabilities, but it is unclear that this
numerical stability is also possible for the construction on the GPU as
kahan summation is inherently sequential.

We also added a couple of utility functions regarding chi-square tests
that we are currently not using.

Our testing suit is getting bigger and now we can finally begin with
implementing the pack method and making some minor adjustments to our
scalar split implementation to get a baseline implementation for the
alias table construction. Finally some progess!