modify prefix partition kernel to output partition as indicies
The partition prefix kernel now writes the partition as indices instead
of values current validation does not directly use this fact directly,
instead it converts is back to the actual partitioned values and
validates them, which is definitly not optimal. Especially because most
of the runtime during validation comes from finding the indices in the
light / heavy partitions, which now is obviously not required anymore,
but well this works reliably at least.

The reason behind this change is that during packing we will require
the partition indices =^(. Future optimizations might find ways around
this.