moving away from std::tuple
We have now moved away from using std::tuple as a memory close type,
because the standard does not define an explicit layout. While changing
our base types, we also modified the partitioning scheme
previously any partition algorithm returned a tuple of
2 spans and a storage, this was sometimes a bit of a hassle to work with
because moving the storage invalidated the spans, therefor we have
introduced a new type wrs::Partition which manages the livetime of the
partitions and ensures that the spans over the individual partitions are
always valid.

We also found a lot of invalid usages of std::span where we often passed
const std::span<T> as a parameter instead of std::span<const T>.
Although we fixed a lot of those invalid usages there are surely a lot
more.