Working baseline YAY!!
Here we implement the first baseline method for construction,
which is composed out of a decoupled mean algorithm to compute the mean
the mean is used to compute a prefix parttion (also decoupled).
Afterwards we run a scalar split and a scalar pack. Here we want to note
that both scalar algorithm are terribly bad GPU implementation in terms
of performance here we where just aiming to get a reference running fast.

Current benchmarks:
10^7 weights -> 4ms