Improves split latency
simply increasing the workgroup size actually already makes a huge
difference i tried around a bit with p-array search strategies to
reduce the search range cooperativly inside of a subgroup but this
didn't lead to a significant improvement, simply because the memory
access speed is dominating here and with p-array search we are not
doing less we are simply doing more of them in parallel which are most
likely not beeing coalessed correctly which leads the same memory
throughput.

I think this is good enough for now we get around 0.048ms if we
split 1024 * 2048 weights into splits with size 8 which is a improvement
of around 5x.

Because we can now actually reduce the split size it might even be
possible to implement a efficient pack step because the about of
serialization required here is reduced by the split size. At least
that's the hope. But still no way that it will perform better than
ITS sampling for sample counts less than the amount of weights.